# docs/api.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CAN-HMI Signal API",
    "description": "Real-time CAN bus signal monitoring and control API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/config": {
      "get": {
        "tags": [
          "Config"
        ],
        "summary": "List all signal configurations",
        "operationId": "list_signal_configs_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SignalConfigResponse"
                  },
                  "type": "array",
                  "title": "Response List Signal Configs Config Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/config/signal/{signal_name}": {
      "get": {
        "tags": [
          "Config"
        ],
        "summary": "Get config for one signal",
        "operationId": "get_signal_config_config_signal__signal_name__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "signal_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Signal Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignalConfigResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/config/processor": {
      "get": {
        "tags": [
          "Config"
        ],
        "summary": "Get processor config",
        "operationId": "get_processor_config_config_processor_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessorConfigResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      },
      "post": {
        "tags": [
          "Config"
        ],
        "summary": "Update processor config",
        "operationId": "update_processor_config_endpoint_config_processor_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProcessorConfigRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessorConfigResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/config/system": {
      "get": {
        "tags": [
          "Config"
        ],
        "summary": "Get system config and field update policy",
        "operationId": "get_system_config_config_system_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Config"
        ],
        "summary": "Patch system config without dropping unrelated fields",
        "operationId": "patch_system_config_config_system_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/config/system/backups": {
      "get": {
        "tags": [
          "Config"
        ],
        "summary": "List fixed-path system config backups",
        "operationId": "list_system_config_backups_config_system_backups_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      },
      "post": {
        "tags": [
          "Config"
        ],
        "summary": "Back up system config",
        "operationId": "backup_system_config_config_system_backups_post",
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/config/system/backups/{backup_id}": {
      "delete": {
        "tags": [
          "Config"
        ],
        "summary": "Delete a system config backup",
        "operationId": "delete_system_config_backup_config_system_backups__backup_id__delete",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "backup_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Backup Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/config/system/backups/{backup_id}/restore": {
      "post": {
        "tags": [
          "Config"
        ],
        "summary": "Restore a system config backup",
        "operationId": "restore_system_config_backup_config_system_backups__backup_id__restore_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "backup_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Backup Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/config/system/reset": {
      "post": {
        "tags": [
          "Config"
        ],
        "summary": "Reset system config from the fixed project template",
        "operationId": "reset_system_config_config_system_reset_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/config/system/reload": {
      "post": {
        "tags": [
          "Config"
        ],
        "summary": "Re-apply live fields from the system config file",
        "operationId": "reload_system_config_config_system_reload_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/config/general": {
      "get": {
        "tags": [
          "Config"
        ],
        "summary": "Get full application config",
        "operationId": "get_general_config_config_general_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Config"
        ],
        "summary": "Patch application config (partial)",
        "operationId": "patch_general_config_config_general_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/config/general/reset": {
      "post": {
        "tags": [
          "Config"
        ],
        "summary": "Reset application config to defaults",
        "operationId": "reset_general_config_config_general_reset_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProcessorConfigResponse": {
        "properties": {
          "max_queue_size": {
            "type": "integer",
            "title": "Max Queue Size",
            "description": "Maximum signal queue size"
          },
          "queue_policy": {
            "type": "string",
            "title": "Queue Policy",
            "description": "Policy when the queue is full: 'drop_oldest' or 'reject'"
          }
        },
        "type": "object",
        "required": [
          "max_queue_size",
          "queue_policy"
        ],
        "title": "ProcessorConfigResponse",
        "description": "Current processor pipeline configuration."
      },
...
      "SignalConfigResponse": {
        "properties": {
          "signal_name": {
            "type": "string",
            "title": "Signal Name",
            "description": "Signal identifier"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit",
            "description": "Measurement unit"
          },
          "min_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Value",
            "description": "Minimum valid value"
          },
          "max_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Value",
            "description": "Maximum valid value"
          },
          "group_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Name",
            "description": "Functional grouping"
          },
          "widget_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Widget Type",
            "description": "Frontend widget type"
          },
          "writable": {
            "type": "boolean",
            "title": "Writable",
            "description": "Whether the signal is writable",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "signal_name"
        ],
        "title": "SignalConfigResponse",
        "description": "Display configuration and metadata for a signal."
      },
...
      "UpdateProcessorConfigRequest": {
        "properties": {
          "max_queue_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Queue Size",
            "description": "New queue size"
          },
          "queue_policy": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "drop_oldest",
                  "reject"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Queue Policy",
            "description": "Handling policy when the queue is full"
          }
        },
        "type": "object",
        "title": "UpdateProcessorConfigRequest",
        "description": "Request to update the processor configuration (PATCH)."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}